buck2: cxx link: don't use `**kwargs`
Summary:
Using `**kwargs` is quite bad because we get zero type-checking.

So, instead, use a record for options. Unfortunately, that is not enough to
get type-checking that works, because:

- Record constructors aren't typed.
- You can't merge fields into a record.

Work around both things by exposing a couple functions to manipulate this new
record.

So, make some functions to fill this in. In time though we should add this to
the Starlark implementation and then we can get rid of those functions.

In theory this would also mean we get to have less duplication in the args
list, but that only really will become true once we get the aforementioned
Starlark-Rust changes.

The reason I'm doing this is to make it easier to add / change the options here
(so that I can enable cache uploads for some binaries).

Reviewed By: shayne-fletcher

Differential Revision: D48116973

fbshipit-source-id: 5f3f94d3fa418180391e3594e76f4c8fa56dc6d6