Introduce LinkStrategy
Summary:
This diff and the previous split LinkStyle into two new types LibOutputStyle and LinkStrategy to better distinguish the meaning of the value where it appears.

A LinkStrategy is mostly about how dependencies are consumed, whether the thing making a shared library wants to consume them as static libs vs shared, but it doesn't
necessarily mean that they will get them in that form (whereas LibOutputStyle will always
map to something that really matches it).

Reviewed By: chatura-atapattu, shayne-fletcher

Differential Revision: D47808802

fbshipit-source-id: ebab2e56ec0ef065afb4e5823656cd0e41c3e2ef