Make shared lib soname opaque
Summary:
Use a new, opaque type for the`soname` field on `SharedLibrary`,
which can hold either a static `str` or an `Artifact` that contains a
build-time generated SONAME, and updated users that require a
static SONAME to enumerate this clearly in the code.

This will allow migrating some rules to support SONAMEs which aren't
known at analysis time (see D55000520).

Reviewed By: shayne-fletcher, igorsugak

Differential Revision: D54989682

fbshipit-source-id: ef3a6b613237d28523f088f0c054e5e34732d6c7