Factor out common package-db flags
We expose all package-dbs of transitive package dependencies of the
current package and use dep-files to achieve recompilation avoidance on
changes to packages that are not transitive dependencies of a given
module in the current package.

This means that the set of package-dbs can be determined once for the
whole package rather than once per module and we can share the generated
package environment file.