Make metadata steps aware of link style and profiling
We want to fully initialize the module graph in the persistent worker
during the metadata step, so that during the compile step no additional
information (i.e. GHC options) except for the unit and module name has
to be provided.
This requires the output style (linking/profiling) to be present among
the flags, and therefore a specialized metadata step for each style,
realized by moving the action into the output style for loops.

Change-Id: Ib68f34db6a0e1a27baa3b2b340d1831300c90e90