haskell generate_target_metadata: show argsfile for failing commands
Previously, `//haskell/tools:generate_target_metadata` would use an
argfile when passing args to `ghc`, but if the `ghc` command failed, it
would print the full invocation (without the argfile) directly [1]. This
is because `generate_target_metadata` would create the argfile in a
temporary directory which was automatically removed when the scope was
exited; therefore, a user attempting to debug a failing invocation would
not be able to determine the actual arguments passed to `ghc`.

To fix this, we declare a new output and pass its path into
`generate_target_metadata` as a new `--use-ghc-args-file-at` argument,
so that `generate_target_metadata` can write a durable argfile which can
be printed later for debugging.

A particularly long `ghc` invocation was printed as 224,413 characters
before this change (1870 lines with a 120-column wide terminal) and 281
characters after this change (3 lines). Much better!

[1]: https://github.com/facebook/buck2-prelude/commit/fa973386098923e85eec5ae1911433749a783bf5

Change-Id: I6a6a69649923ff9305388029cde54fb7e8edd49d