Improve missing library dep error
In some circumstances (as of yet unclear exactly when this happens -
possibly related to module re-exports in a package's cabal file), we
run into this error before invoking GHC. Unfortunately, when this does
happen, the error is not at all clear - in particular it's not clear
which module/target the error arises from. This change prints the
module. Printing the target name as well would be ideal, but I wasn't
immediately able to figure out how to do that. I'm happy to take another
pass if we think that would be useful though.

Change-Id: I34708caf8e37a3772a6373d2f1582642719d8178