Python wrapper of ghc-pkg register operation and fix incorrect GHC_PACKAGE_PATH use
Replace _REGISTER_PACKAGE inlined shell script by a python script,
and use a unified _register_package_conf at two use sites.

It turns out that GHC_PACKAGE_PATH should not be used for a ghc-pkg operation,
so the env var is removed and the corresponding dependencies db_deps are removed.

Change-Id: I1d1d1315a6148b59b3bb0429aa8799626bf3eb09