GhcLinkableInfo for providing dynamic linker flags for extra libraries
This introduces GhcLinkableInfo which has dynamic
ExtraGhcLinkerFlagsInfo, so that linker parameters can be dynamically
obtained from additional extra steps. For example, We can use this
dynamic capability to find out RPATH info for nix-provided external
libraries.

The actual library is wrapped in the GhcLinkableInfo as
MergedLinkInfo, so transitive dependency sets are calculated in the
same way as before.

Only linker-time flags are additionally provided with proper wrapping
with `-optl` for GHC as a linker. The options are also encoded in
package DB explicitly as `ld-options`.

Change-Id: I26fc39f0a977b68d5caeea3e9ef239b40bbe2a2b