Fix -package parameters for compile_module
The package list set by -package for each module build should contain
all the packages needed for transitive dependency modules in the the
current unit of a given module under compilation.

This is especially relevant to TH with extra C dependencies because
GHC can only look up relevant info of C deps through package DB.
So when categorizing package deps, we compute that information and
add them to library_deps, so that they can be added to -package.

Change-Id: I01ae145e90ff51cbb55baf8b29f963b46d079c4b